Move mandatory initialisation into framework
